+++ VTK/Rendering/vtkImageMapper.cxx
@@ -157,7 +157,7 @@ void vtkImageMapper::RenderStart(vtkViewport* viewport, vtkActor2D* actor)
     int *pos = actor->GetPositionCoordinate()->GetComputedViewportValue(viewport);
 
     // Get the viewport coordinates
-    float vCoords[4];
+    double vCoords[4];
     vCoords[0] = 0.0;
     vCoords[1] = 0.0;
     vCoords[2] = 1.0;
