+++ VTK/Rendering/vtkInteractorStyleTerrain.cxx
@@ -287,7 +287,7 @@ void vtkInteractorStyleTerrain::Dolly()
 
   vtkRenderWindowInteractor *rwi = this->Interactor;
   vtkCamera *camera = this->CurrentRenderer->GetActiveCamera();
-  float *center = this->CurrentRenderer->GetCenter();
+  double *center = this->CurrentRenderer->GetCenter();
 
   int dy = rwi->GetEventPosition()[1] - rwi->GetLastEventPosition()[1];
   double dyf = this->MotionFactor * (double)(dy) / (double)(center[1]);
