+++ VTK/Rendering/vtkWorldPointPicker.cxx
@@ -33,8 +33,8 @@ int vtkWorldPointPicker::Pick(double selectionX, double selectionY,
 {
   vtkCamera *camera;
   double cameraFP[4];
-  float display[3], *world;
-  float *displayCoord;
+  double display[3], *world;
+  double *displayCoord;
   double z;
 
   // Initialize the picking process
